requirements.txt
cv2box
apstone
numpy
onnxruntime
numba==0.54.1
nvidia_tensorrt==8.0.0.3
onnx==1.10.1
scipy==1.6.2
torchvision==0.9.1+cu111
scikit_image==0.18.3
tqdm==4.61.1
torch==1.8.1+cu111
pycuda==2021.1
matplotlib==3.4.3
Pillow==8.4.0
